perf: code split to chat box
